consistently infer the null state from the map function
There were a few places where we were assuming that the null state
would be the empty vector, but that's not necessarily the case with
arbitrary map and compose specifications for parsers.

The parser's state must always be initialized to the null state
inferred from the parser's `map` function, and difference from this
null state is what indicates that the parser is in the middle of
parsing.

We no longer need the guard check for nonempty state at the accept
stage, as the state would only be accepted if parsing was in progress
to begin with, which is *defined* as the parser having non-null state.