check abort predicate at the start of each parsing step
Formerly it was being checked right before checking the accept
predicate, at the end of the parsing step. But this leads to an
infinite loop in the case where a key sequence to repeat a previous
command occurs when parsing is already in progress, even if that key
sequence has been explicitly excluded from parsing via an abort
condition.

For instance, in `2 .` to repeat the last command two times, if the
previous command was `x`, then this ends up parsing `2.x`. If we
configure it to abort on encountering `.`, it doesn't in this case
since the `x` was appended in a nested command before it gets to the
outer post-command stage (and I suppose that also ends up overriding
`this-command-keys` in the outer command?). Then, the logic to repeat
"twice" attempts to repeat the command once again, this time evidently
replaying `2 . x`. But since `.` once again replays `2 . x`, this ends
up in an infinite loop of replaying (and also parsing) `2 . 2 . 2 .
...`.

Checking the abort predicate before command execution avoids this
problem.