ability to indicate pairwise parser subscriptions
Formerly, to implement higher-level parsers that used tokens parsed by
lower-level parsers, it was necessary to hook them up manually after
subscribing to the lower-level parser in a function. We now provide a
function to abstract this so that we only need to indicate a
subscription relationship between two parsers, and it will take care
of subscribing to the former and feeding the tokens to the latter.