distinguish key vectors as the primitive, and `key` as a key string
This allows us to use explicit key vectors as the primitive
representation for the evaluator, but still support convenient string
representations in common cases.

The translation from key strings to key vectors is straightforward.
It's unclear whether there are cases that cannot unambiguously be
expressed as strings (necessitating the vector representation in
Emacs). In any case, using numeric vectors as the lowest level feels
safest, and this way, we don't have to compromise on convenience,
either.