add "insertion" as a primitive operation in the DSL
Formerly, key vectors were the only primitive. This adds insertion as
a second primitive.

Technically, we could just do this using the fallback lambda case, and
that may be more appropriate if we consider that Mantra deals
explicitly only with key sequences, i.e., the syntax of Emacs.

On the other hand, having `insert` as an explicit form would likely
read better in parsed mantras, for "rich repeat" (e.g., in Symex)
purposes. Arguably, insertion is a special case where that syntax is
intended to be unevaluated, where the syntax *is* the semantics, in a
way. So maybe it's OK to have this special form, without it implying a
Pandora's box of other primitive operations.