do parsing entirely in the post-command stage
This transfers the responsibility of knowing the pre-command state (if
that's necessary) to the parser itself. For instance, a parser used in
a certain context should be *subscribed* and active only in that
context, so that it can assume the context implicitly without it
needing to be part of the start condition. Care should be taken to
unsubscribe the parser outside of that context.

This frees us to move the start condition into the post-command stage
along with the rest of the parsing steps, as we no longer need
pre-command awareness. This makes primitive parsing uniform with
higher-level parsing, and opens the door to further streamlining of
the implementation in light of this uniformity.