"finish" parsed state before publishing
This allows us to decouple the composition of the parsed data from the
final type produced. For instance, we could parse by accumulating a
list, but ultimately produce a custom type from the parsed list. This
would otherwise need to be done by the client either via an extra
indirection at the pub/sub level, or by defining composition of their
custom type and parsing this custom type all the way through, which
might incur a performance cost. It seems more properly to belong to
the parser itself, as a way to parse a nonlocal, "gestalt" result,
imbuing the finished state with properties not inherent in any
component.