add `move` form to express point motion
There are cases where there is point motion without an explicit
attendant key sequence. For example, a command that inserts text in
the buffer and also moves point could not be captured either purely as
a key sequence or purely as an insertion. In such cases, we need an
explicit `move` form to be able to decompose the complex action into
separate insertion and motion components.