ability to express (e.g., interactive) commands in mantras
Commands are evaluated simply by evaluating the command referred to by
them. Being able to express command execution in the DSL allows us to
compose mantras whose effects need not be reducible to insertions and
deletions. Key sequence-based mantras also achieve this, but parsing
commands directly allows them to be replayed in a keybindings-agnostic
way.

We already do support arbitrary lambdas as mantras, but supporting
named commands makes this familiar case first class, so that parsing
them would not require a wrapping lambda.